+++ packages/language-support/src/highlighting/syntaxValidation/semanticAnalysisWrapper.ts
@@ -2,6 +2,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
+import { DiagnosticSeverity } from 'vscode-languageserver-types';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import { semanticAnalysis } from './semanticAnalysis';
@@ -12,6 +13,7 @@ export interface SemanticAnalysisResult {
 }
 
 export interface SemanticAnalysisElement {
+  severity: DiagnosticSeverity;
   message: string;
   position: {
     offset: number;
@@ -29,7 +31,27 @@ export function doSemanticAnalysis(query: string): SemanticAnalysisResult {
     const errors: SemanticAnalysisElement[] = semanticErrorsResult.$errors.data;
     const notifications: SemanticAnalysisElement[] =
       semanticErrorsResult.$notifications.data;
-    return { errors: errors, notifications: notifications };
+
+    return {
+      errors: errors.map(({ message, position }) => ({
+        severity: DiagnosticSeverity.Error,
+        message,
+        position: {
+          column: position.column,
+          line: position.line,
+          offset: position.offset,
+        },
+      })),
+      notifications: notifications.map(({ message, position }) => ({
+        severity: DiagnosticSeverity.Warning,
+        message,
+        position: {
+          column: position.column,
+          line: position.line,
+          offset: position.offset,
+        },
+      })),
+    };
   } catch (e) {
     /* Ignores exceptions if they happen calling the semantic analysis. Should not happen but this is just defensive in case it did */
     return { errors: [], notifications: [] };
