+++ packages/language-support/tsconfig.json
@@ -5,7 +5,6 @@
   "compilerOptions": {
     "outDir": "out",
     "declaration": true,
-    "allowJs": true,
-    "module": "ESNext"
+    "allowJs": true
   }
 }
