+++ packages/react-codemirror/src/e2e_tests/syntax-validation.spec.tsx
@@ -154,3 +154,5 @@ test('Validation errors are correctly overlapped', async ({ page, mount }) => {
     "Invalid input. '-1' is not a valid value. Must be a positive integer",
   );
 });
+
+// TODO test semantic analayssi here as well
